revise ttirtottkernel into ttirtottmetal and ttirtottkernel dialect conversions - ERASES OLD TTMETAL LOWERING
