+++ include/ttmlir/Conversion/TTIRToTTMetal/TTIRToTTMetal.h
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+#include "ttmlir/Conversion/TTIRToTTMetal/TTIRToTTKernel.h"
+
 #ifndef TTMLIR_CONVERSION_TTIRTOTTMETAL_TTIRTOTTMETAL_H
 #define TTMLIR_CONVERSION_TTIRTOTTMETAL_TTIRTOTTMETAL_H
 
@@ -15,7 +17,6 @@ void populateTTIRToTTMetalPatterns(MLIRContext *ctx,
                                    TypeConverter &typeConverter);
 
 std::unique_ptr<OperationPass<ModuleOp>> createConvertTTIRToTTMetalPass();
-std::unique_ptr<OperationPass<ModuleOp>> createConvertTTIRToTTKernelPass();
 
 } // namespace mlir::tt
